moved the examles to a separate file and renamed the interface to Node
