+++ shrdlite.html
@@ -16,6 +16,7 @@
   <script type="text/javascript" src="Interpreter.js"></script>
   <script type="text/javascript" src="shrdlite-html.js"></script>
   <script type="text/javascript" src="AStar.js"></script>
+  <script type="text/javascript" src="AStarExample.js"></script>
   <!-- <script type="text/javascript" src="shrdlite-ajax.js"></script> -->
 </head>
 <body>
